refactor: aircraft display into self contained widgets within `tangram_jet1090` plugin
- core `MapApi` now owns and provides the leaflet instance
- render markers from the reactive `StateApi`
- removed hacky logic to detect which aircraft is selected
- use `lit-html` properly for svg and tooltips to avoid XSS
  (not using vue `createApp` for perf)
