+++ packages/tangram/index.html
@@ -16,17 +16,14 @@
       rel="stylesheet"
       href="https://unpkg.com/bootstrap@5.1.3/dist/css/bootstrap.min.css"
     />
-    <link
-      rel="stylesheet"
-      href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css"
-      crossorigin=""
-    />
     <!-- we externalise vue esm so plugins dont have to vendor their own -->
     <!-- this results in a larger (initial) bundle size but is well worth it -->
     <script type="importmap">
       {
         "imports": {
-          "vue": "/vue.esm-browser.prod.js"
+          "vue": "/vue.esm-browser.prod.js",
+          "leaflet": "/leaflet-src.esm.js",
+          "lit-html": "/lit-html.js"
         }
       }
     </script>
