--- packages/tangram_jet1090/src/tangram_jet1090/Jet1090Widget.vue
@@ -1,5 +1,0 @@
-<template>
-  <div>
-    <h3>@open-aviation/tangram-jet1090</h3>
-  </div>
-</template>
