+++ packages/tangram_jet1090/src/tangram_jet1090/PlanePath.ts
@@ -1,0 +1,806 @@
+export interface IconProps {
+  x: number;
+  y: number;
+  scale: number;
+  rotcorr: number;
+  path: string;
+  ofX?: number;
+  ofY?: number;
+}
+
+export interface AcSymbol {
+  id: string;
+  sym: string;
+}
+
+let acImagePathsR: Record<string, IconProps> = {
+  // triangle
+  $0: { x: 12, y: 12, scale: 1, rotcorr: 180, path: "M0,0 6,12 12,0z" },
+  // small aircraft
+  $1: {
+    x: 16,
+    y: 17,
+    scale: 1,
+    rotcorr: 180,
+    path: "M8,17 9,16 9,12 16,11 16,8 9,7 9,3 11,1 11,0 8,1 5,0 5,1 7,3 7,7 0,8 0,11 7,12 7,16 8,17z"
+  },
+  // two propeller aircraft
+  $2: {
+    x: 26,
+    y: 18,
+    scale: 1,
+    rotcorr: 180,
+    path: "M13,18 14,17 14,17 15,14 17,14 17,16 18,17 19,16 19,14 25,14 26,13 26,12 26,12 25,11 19,10 19,9 18,9 18,10 15,9 14,6 14,2 18,1 18,0 13,0 8,0 8,1 12,2 12,6 11,9 8,10 8,9 7,9 7,10 1,11 0,12 0,12 0,13 1,14 7,14 7,16 8,17 9,16 9,14 11,14 12,17 12,17 13,18z"
+  },
+  // jet with two turbines at back (Learjet or so)
+  $3: {
+    x: 21,
+    y: 24,
+    scale: 1,
+    rotcorr: 180,
+    path: "M11,24 12,23 12,14 20,10 21,8 13,10 12,9 12,7 14,7 13,4 12,4 12,3 14,2 15,0 12,1 11,0 10,1 7,0 8,2 10,3 10,4 9,4 8,7 10,7 10,9 9,10 0,8 1,10 10,14 10,23 11,24z"
+  },
+  // A320, B737
+  $4: {
+    x: 24,
+    y: 28,
+    scale: 1,
+    rotcorr: 180,
+    path: "M12,28 13,26 14,23 14,18 16,17 16,19 18,19 18,16 24,11 24,10 23,10 15,13 14,13 13,6 13,5 16,2 16,0 13,2 12,0 11,2 8,0 8,2 11,5 11,6 10,13 9,13 1,10 0,10 0,11 6,16 6,19 8,19 8,17 10,18 10,23 11,26 12,28z"
+  },
+  // B747, A380
+  $5: {
+    x: 30,
+    y: 31,
+    scale: 1,
+    rotcorr: 180,
+    path: "M15,31 17,28 17,22 20,20 20,22 22,22 22,19 25,17 25,19 27,19 27,16 30,13 30,11 19,14 17,14 17,7 16,5 20,2 20,0 16,1 15,0 14,1 10,0 10,2 14,5 13,7 13,14 11,14 0,11 0,13 3,16 3,19 5,19 5,17 8,19 8,22 10,22 10,20 13,22 13,28 15,31z"
+  },
+  // Fighter
+  $6: {
+    x: 16,
+    y: 20,
+    scale: 1,
+    rotcorr: 180,
+    path: "M8,20 9,18 9,14 10,12 10,11 16,8 15,6 10,5 10,3 12,2 12,1 11,0 10,0 9,2 8,1 7,2 6,0 5,0 4,1 4,2 6,3 6,5 1,6 0,8 6,11 6,12 7,14 7,18 8,20z"
+  },
+  // Helicopter
+  $7: {
+    x: 20,
+    y: 23,
+    scale: 1,
+    rotcorr: 180,
+    path: "M10,20 11,20 11,20 12,17 19,23 20,22 13,15 13,13 20,6 19,5 12,11 11,9 11,4 12,4 13,2 11,2 11,1 10,0 9,1 9,2 7,2 8,4 9,4 9,9 8,11 1,5 0,6 7,13 7,15 0,22 1,23 8,17 9,20 10,20z"
+  },
+  // jet with two turbines at back (smaller than $3)
+  $8: {
+    x: 16,
+    y: 20,
+    scale: 1,
+    rotcorr: 180,
+    path: "M8,20 9,18 9,12 16,7 16,6 9,7 10,5 10,4 9,3 10,2 11,1 11,0 8,1 5,0 5,1 6,2 7,3 6,4 6,5 7,7 0,6 0,7 7,12 7,18 8,20z"
+  },
+  // small jet with 4 turbines on wings, Avro RJ70
+  $9: {
+    x: 22,
+    y: 18,
+    scale: 1,
+    rotcorr: 180,
+    path: "M11,18 12,17 13,14 13,12 15,12 15,13 17,13 17,11 19,10 19,11 21,11 21,9 22,8 22,6 14,8 13,8 13,7 12,2 12,2 15,1 15,0 12,0 11,0 10,0 7,0 7,1 10,2 10,2 9,7 9,8 8,8 0,6 0,8 1,9 1,11 3,11 3,10 5,11 5,13 7,13 7,12 9,12 9,14 10,17 11,18z"
+  },
+  // Osprey (takeoff)
+  $A: {
+    x: 30,
+    y: 21,
+    scale: 1,
+    rotcorr: 180,
+    path: "M14,20 16,20 17,18 17,14 21,14 18,20 19,21 20,20 22,14 23,13 23,13 29,14 29,14 30,13 29,12 23,12 22,11 22,5 21,4 20,5 21,11 17,11 17,9 16,6 16,3 18,3 18,0 12,0 12,3 14,3 14,6 13,9 13,11 9,11 10,5 9,4 8,5 8,11 7,12 2,12 1,12 0,13 1,14 7,13 8,14 8,14 10,20 11,21 12,20 9,14 13,14 13,18 14,20z"
+  },
+  // Osprey (flight)
+  $A2: {
+    x: 27,
+    y: 20,
+    scale: 1,
+    rotcorr: 180,
+    path: "M14,20 15,19 15,15 16,14 19,14 20,15 16,17 20,17 20,18 21,18 21,17 27,17 21,15 22,14 21,11 21,10 20,10 20,11 16,11 16,10 15,6 15,3 16,3 16,0 11,0 11,3 12,3 12,6 11,10 11,11 7,11 7,10 6,10 5,14 6,15 0,17 6,17 6,18 7,18 7,17 11,17 7,15 8,14 11,14 12,15 12,19 13,20 14,20z"
+  },
+  // Transall
+  $B: {
+    x: 21,
+    y: 20,
+    scale: 1,
+    rotcorr: 180,
+    path: "M10,20 12,19 12,17 12,13 15,13 15,13 15,13 15,15 17,15 17,13 19,13 19,13 21,12 21,10 19,9 12,9 12,4 12,3 14,2 14,0 12,0 10,0 9,0 7,0 7,2 9,3 9,4 9,9 2,9 0,10 0,12 2,13 2,13 4,13 4,15 6,15 6,13 7,13 7,13 9,13 9,17 9,19 10,20z"
+  },
+  // Hercules
+  $C: {
+    x: 26,
+    y: 24,
+    scale: 1,
+    rotcorr: 180,
+    path: "M13,24 14,23 15,22 15,16 17,16 17,18 19,18 19,16 21,16 21,18 23,18 23,16 26,15 26,13 23,12 15,12 15,4 14,3 17,2 17,0 14,0 13,0 12,0 9,0 9,2 12,3 11,4 11,12 3,12 0,13 0,15 3,16 3,18 5,18 5,16 7,16 7,18 9,18 9,16 11,16 11,22 12,23 13,24z"
+  },
+  // ???
+  $D: {
+    x: 28,
+    y: 25,
+    scale: 1,
+    rotcorr: 180,
+    path: "M14,25 15,22 15,14 17,14 17,17 18,18 19,17 19,14 28,14 28,12 15,10 15,3 19,1 19,0 9,0 9,1 13,3 13,10 0,12 0,14 9,14 9,17 10,18 11,17 11,14 13,14 13,22 14,25z"
+  },
+  // Military Helicopter
+  $E: {
+    x: 18,
+    y: 30,
+    scale: 1,
+    rotcorr: 180,
+    path: "M9,28 10,28 10,28 11,24 17,30 18,29 11,20 11,19 18,10 17,9 11,15 11,9 10,6 10,6 10,4 11,3 10,1 9,0 8,0 5,0 5,3 8,4 8,6 7,9 7,15 1,9 0,10 7,19 7,20 0,29 1,30 7,24 8,28 9,28z"
+  },
+  // Glider:
+  $SF25: {
+    x: 10,
+    y: 5,
+    scale: 1.6,
+    rotcorr: 0,
+    path: "m 9.9426045,9.250116 c 0,-0.105041 -0.019648,-0.245689 -0.043664,-0.312551 C 9.8649535,8.842935 9.8108295,8.371017 9.8045435,8.114493 9.803882,8.087283 9.5539473,8.079763 8.6498441,8.079763 H 7.4959966 l 0.00783,-0.325223 0.00783,-0.325224 1.10985,-0.147948 C 9.2319245,7.199998 9.737399,7.127377 9.7447838,7.119992 9.7888508,7.075922 9.5909678,4.190505 9.5404112,4.139948 9.5329512,4.132448 7.3856524,3.912321 4.7686375,3.650689 l -4.75820654,-0.475695 -0.007587,-0.492542 -0.007587,-0.4925422 0.0770424,-3.051e-4 c 0.0423733,-1.678e-4 1.88358804,-0.031286 4.09158834,-0.069151 2.208,-0.037865 4.2793269,-0.068983 4.6029483,-0.069151 l 0.5883962,-3.018e-4 0.011456,-0.2430953 c 0.0063,-0.1337025 0.027051,-0.4368771 0.046111,-0.6737214 0.03932,-0.48861447 0.054908,-0.52879147 0.2856822,-0.73627707 0.080507,-0.072383 0.1693598,-0.1869829 0.205571,-0.2651414 0.07983,-0.1723046 0.099911,-0.1718062 0.1878472,0.00466 0.03901,0.078275 0.121558,0.1856619 0.183447,0.238637 0.271082,0.2320359 0.304028,0.3651824 0.354301,1.43184117 l 0.01146,0.2430953 0.588402,3.043e-4 c 0.323622,1.678e-4 2.394949,0.031286 4.602949,0.069151 2.208,0.037865 4.049215,0.068983 4.091588,0.069151 l 0.07704,3.043e-4 -0.0076,0.4925422 -0.0076,0.492542 -4.758208,0.475695 c -2.617015,0.261632 -4.764313,0.481799 -4.771774,0.489259 -0.04986,0.04986 -0.248672,2.935744 -0.20524,2.979177 0.0069,0.0069 0.512383,0.07909 1.123278,0.160399 l 1.110717,0.147838 0.0078,0.326201 0.0078,0.3262 h -1.1499 -1.149978 l -0.01752,0.145857 c -0.0096,0.08022 -0.01793,0.214619 -0.01842,0.29866 -5.02e-4,0.08404 -0.0136,0.202811 -0.02912,0.263932 -0.01552,0.06112 -0.03153,0.126395 -0.03558,0.145054 -0.0041,0.01866 -0.01606,0.04262 -0.02669,0.05325 -0.01063,0.01063 -0.01933,0.117258 -0.01933,0.236954 0,0.199106 -0.0047,0.217628 -0.05556,0.217628 -0.050176,0 -0.055565,-0.01852 -0.055565,-0.190983 z"
+  },
+  $SI2: {
+    x: 26,
+    y: 29,
+    scale: 0.9,
+    rotcorr: 0,
+    path: "m 19.964368,1071.9917 c 0,-0.1479 -0.708303,-0.2399 -1.847142,-0.2399 -1.25149,0 -1.900887,-0.093 -2.013803,-0.2879 -0.215611,-0.3725 -0.215611,-1.9306 0,-2.303 0.110404,-0.1908 0.708126,-0.2879 1.770757,-0.2877 l 1.604097,2e-4 -0.04023,-4.7805 -0.04023,-4.7804 -7.667766,-0.1393 c -8.938858,-0.1622 -10.6152712,-0.2754 -10.88130821,-0.735 -0.2481701,-0.4287 -0.24320028,-2.2209 0.007396,-2.6539 0.23995191,-0.4144 4.34617251,-0.8175 8.37957011,-0.8224 l 2.6734951,0 0,-0.7198 c 0,-0.6397 0.04629,-0.7196 0.416649,-0.7196 0.370354,0 0.416647,0.08 0.416647,0.7196 0,0.7198 0,0.7198 0.833298,0.7198 0.833297,0 0.833297,0 0.833297,-0.7198 0,-0.6397 0.04629,-0.7196 0.416649,-0.7196 0.370354,0 0.416649,0.08 0.416649,0.7196 l 0,0.7198 2.152684,0 2.152685,0 0,-0.6621 c 0,-0.402 0.109014,-0.7796 0.277764,-0.9616 0.39401,-0.4251 0.833298,0.082 0.833298,0.9616 l 0,0.6621 2.152685,0 2.152684,0 0,-0.7198 c 0,-0.6397 0.04629,-0.7196 0.416649,-0.7196 0.370354,0 0.416649,0.08 0.416649,0.7196 0,0.7198 0,0.7198 0.833296,0.7198 0.833298,0 0.833298,0 0.833298,-0.7198 0,-0.6397 0.04629,-0.7196 0.416648,-0.7196 0.370355,0 0.416649,0.08 0.416649,0.7196 l 0,0.7198 2.673494,0 c 4.033398,0.01 8.13962,0.408 8.379571,0.8224 0.250558,0.433 0.255527,2.2252 0.0073,2.6539 -0.266036,0.4596 -1.942448,0.5728 -10.881306,0.735 l -7.667767,0.1393 -0.04023,4.7804 -0.04023,4.7805 1.604097,-2e-4 c 1.06263,-2e-4 1.660353,0.097 1.770756,0.2877 0.215612,0.3724 0.215612,1.9305 0,2.303 -0.112916,0.1951 -0.762312,0.2879 -2.013802,0.2879 -1.138839,0 -1.847141,0.092 -1.847141,0.2399 0,0.132 -0.06249,0.24 -0.138882,0.24 -0.07639,0 -0.138884,-0.108 -0.138884,-0.24 z"
+  },
+  $AWACS: {
+    x: 26,
+    y: 29,
+    scale: 0.82,
+    rotcorr: 0,
+    path: "m 14.343596,1092.446 c -0.625928,-1.9855 0.923503,-3.2681 2.357528,-4.3114 0.638047,-0.911 2.558375,-1.4137 2.29187,-2.6824 -0.362274,-1.7564 0.05029,-3.8316 -1.945007,-4.6377 -1.636357,-1.7445 -1.31938,-4.7335 0.546742,-6.1856 -1.745001,0.2461 -3.609901,0.6206 -5.302719,1.249 -3.5415964,1.6938 -6.9829162,3.632 -10.5785055,5.1941 -0.5439413,-2.14 1.2146542,-3.7803 2.8415081,-4.8678 0.5261676,-0.6676 1.8896692,-0.9291 1.6238653,-1.9709 0.23028,-1.226 -0.6050837,-3.2618 0.47332,-3.8719 2.6153309,-0.9549 -0.026897,4.4345 1.9627108,2.5743 0.6938338,-0.8966 2.4992633,-1.3501 2.4085743,-2.6254 0.211105,-1.2601 -0.815967,-3.5113 0.535436,-3.9942 2.490841,-0.7901 -0.244444,4.8383 1.951076,2.6156 1.608072,-1.2923 3.214227,-2.5869 4.817539,-3.8851 0.08059,-2.9649 -0.196262,-5.961 0.261681,-8.8989 0.412822,-0.9973 0.517997,-4.4527 1.939053,-3.23 1.221951,2.7783 1.301796,5.8309 1.255964,8.8128 -0.02924,1.8951 -0.340617,4.0059 1.787657,4.8661 1.290079,0.8787 2.933003,2.6509 4.028181,2.9569 -0.06602,-1.3591 -0.605246,-4.008 1.619196,-3.1183 0.291731,1.5329 -0.113716,3.114 0.0099,4.6766 1.096138,0.6701 2.341176,2.1783 3.286665,2.3533 -0.311845,-1.3348 -0.553736,-3.8948 1.60534,-2.949 0.275013,1.4619 -0.09839,2.9696 -0.05541,4.4571 1.362156,1.1016 2.860751,2.0858 4.037065,3.3829 0.39777,1.2232 0.467414,3.6305 -1.233984,1.9739 -3.181761,-1.5474 -6.287606,-3.2706 -9.532607,-4.6755 -1.364436,-0.2822 -3.557771,-1.036 -4.488664,-0.9268 2.138134,1.6847 2.010353,5.3437 -0.352372,6.7679 -0.904477,0.2105 -0.572176,1.215 -0.75079,1.8975 -0.781025,2.0295 0.104268,3.6169 1.922096,4.6683 1.341759,1.1313 3.212014,2.4444 2.427988,4.4569 -1.805242,-0.1579 -4.316554,-2.0576 -5.585939,-1.4884 -0.33527,1.5767 -0.467274,-1.2308 -1.385651,-0.1469 -1.606227,0.4311 -3.160352,1.3124 -4.779356,1.563 z"
+  },
+  $A320: {
+    x: 20,
+    y: 20,
+    scale: 0.75,
+    rotcorr: 0,
+    path: "m 32,4.8 0.3,0.1 0.8,1.2 0.9,3.1 0.6,3.1 0.5,5.2 0,6.9 3.6,2.7 -0.2,-1.5 0,-1.5 0.1,-1.2 0.2,-0.4 2.8,0 0.2,0.4 0.1,1.2 0,1.5 -0.4,2.2 -0.3,0 -0.1,0.9 14,7.7 0.4,0.4 0.2,0.7 0,1.5 -8.3,-2.4 -0.1,0.4 -0.2,0.2 -0.2,-0.2 -0.1,-0.5 -3.7,-1.1 -0.1,0.4 -0.2,0.2 -0.2,-0.2 -0.1,-0.5 -2,-0.6 0,1 -0.1,0.5 -0.1,0.1 -0.4,0 -0.1,-0.1 -0.1,-0.5 0,-1.3 -4.6,0 0,7.7 -0.1,2.8 -0.6,3.7 -0.6,3 0.1,0.4 8.5,5.9 0,1.8 -10,-2.5 -0.2,0.7 -0.1,0.8 -0.1,0.4 -0.1,-0.4 -0.1,-0.8 -0.2,-0.7 -10,2.5 0,-1.8 8.5,-5.9 0.1,-0.4 -0.6,-3 -0.6,-3.7 -0.1,-2.8 0,-7.7 -4.6,0 0,1.3 -0.1,0.5 -0.1,0.1 -0.4,0 -0.1,-0.1 -0.1,-0.5 0,-1 -2,0.6 -0.1,0.5 -0.2,0.2 -0.2,-0.2 -0.1,-0.4 -3.7,1.1 -0.1,0.5 -0.2,0.2 -0.2,-0.2 -0.1,-0.4 -8.3,2.4 0,-1.5 0.2,-0.7 0.4,-0.4 14,-7.7 -0.1,-0.9 -0.3,0 -0.4,-2.2 0,-1.5 0.1,-1.2 0.2,-0.4 2.8,0 0.2,0.4 0.1,1.6 0,1.1 -0.2,1.5 3.6,-2.8 0,-6.8 0.5,-5.2 0.6,-3.1 1,-3.1 0.7,-1.2 z"
+  },
+  $A330: {
+    x: 20,
+    y: 20,
+    scale: 0.7,
+    rotcorr: 0,
+    path: "m 32,0.1 0.1,0 0.2,0.1 0.4,0.3 0.5,0.8 0.7,1.7 0.5,2 0.2,2.6 0,15.4 0.6,0.8 3.9,2.7 -0.2,-1.1 0,-1.2 0.2,-1.1 0.1,-0.2 2.3,0 0.1,0.2 0.2,1.1 0,1.2 -0.2,1.2 -0.3,0 -0.1,1.2 15.6,10.6 2.1,3.1 0,1 -2.1,-1.6 -7.8,-3.2 -0.1,0.4 -0.1,-0.5 -4.2,-1.4 -0.1,0.4 -0.1,-0.5 -4.4,-1.4 -1,-0.2 -0.2,0.5 -0.2,-0.5 -4,-0.4 0,14.2 -0.1,2.6 -0.4,4 0.5,0.7 7.2,5.8 0,1.6 -8.9,-2.3 -0.7,1.5 -0.1,1.3 -0.2,0 -0.1,-1.3 -0.7,-1.5 -8.9,2.3 0,-1.6 7.2,-5.8 0.5,-0.7 -0.4,-4 -0.1,-2.6 0,-14.2 -4,0.4 -0.2,0.5 -0.2,-0.5 -1,0.2 -4.4,1.4 -0.1,0.5 -0.1,-0.4 -4.2,1.4 -0.1,0.5 -0.1,-0.4 -7.8,3.2 -2.1,1.6 0,-1 2.1,-3.1 15.6,-10.6 -0.1,-1.2 -0.3,0 -0.2,-1.2 0,-1.2 0.2,-1.1 0.1,-0.2 2.3,0 0.1,0.2 0.2,1.1 0,1.2 -0.2,1.1 3.9,-2.7 0.6,-0.8 0,-15.4 0.2,-2.6 0.5,-2 0.7,-1.7 0.5,-0.8 0.4,-0.3 0.2,-0.1 z"
+  },
+  $A340: {
+    x: 20,
+    y: 20,
+    scale: 0.7,
+    rotcorr: 0,
+    path: "m 32,0.2 0.1,0 0.2,0.1 0.3,0.3 0.5,0.8 0.7,1.7 0.5,2 0.3,2.5 0,15 0.3,0.6 0.5,0.6 3.3,2.2 -0.2,-1.1 0,-1.6 0.2,-1.4 0.1,-0.2 3,0 0.1,0.2 0.2,1.4 0,1.6 -0.2,1.4 -0.4,0 -0.4,1.3 16.4,10.8 0.2,0.5 0.1,1.5 -9,-3.1 -0.1,0.6 -0.1,0.2 -0.1,-0.2 -0.1,-0.7 -4.2,-1.4 -0.1,0.6 -0.1,0.2 -0.1,-0.2 -0.1,-0.7 -2.5,-0.8 -1.8,0 -0.1,0.7 -0.1,0.2 -0.1,-0.2 -0.1,-0.7 -4.5,0 0,14.9 -0.2,1.9 -0.7,3.8 7.6,6.2 0.1,0.7 0,0.7 -0.1,0.7 -8.5,-3.1 -0.2,0.4 -0.5,1.8 -0.2,0 -0.5,-1.8 -0.2,-0.4 -8.5,3.1 -0.1,-0.7 0,-0.7 0.1,-0.7 7.6,-6.2 -0.7,-3.8 -0.2,-1.9 0,-14.9 -4.5,0 -0.1,0.7 -0.1,0.2 -0.1,-0.2 -0.1,-0.7 -1.8,0 -2.5,0.8 -0.1,0.7 -0.1,0.2 -0.1,-0.2 -0.1,-0.6 -4.2,1.4 -0.1,0.7 -0.1,0.2 -0.1,-0.2 -0.1,-0.6 -9,3.1 0.1,-1.5 0.2,-0.5 16.4,-10.8 -0.4,-1.3 -0.4,0 -0.2,-1.4 0,-1.6 0.2,-1.4 0.1,-0.2 3,0 0.1,0.2 0.2,1.4 0,1.6 -0.2,1.1 3.3,-2.2 0.5,-0.6 0.3,-0.6 0,-15 0.3,-2.5 0.5,-2 0.7,-1.7 0.5,-0.8 0.3,-0.3 0.2,-0.1 z"
+  },
+  $A380: {
+    x: 20,
+    y: 20,
+    scale: 0.86,
+    rotcorr: 0,
+    path: "m 20.170952,35.711605 c 0.240955,-0.452026 0.835455,-1.57222 0.835455,-1.57222 l 6.03507,2.126063 1.056047,-1.025537 -6.496618,-6.438864 0.883688,-9.623753 16.799383,6.948921 0.461547,-2.906545 -6.354074,-4.307448 0.109254,-3.008361 -1.776828,-1.083222 -0.608121,2.103338 -2.727027,-1.738882 0.337976,-3.132209 -1.375745,-1.158211 -1.689327,2.551609 -3.321597,-3.620931 0.01293,-3.9857346 -1.062159,-4.86699002 -1.125964,-0.70583829 h -0.08889 l -1.125965,0.70583829 -1.060843,4.86699002 0.01153,3.9857346 -3.322291,3.620931 -1.687937,-2.551609 -1.37505,1.158211 0.335961,3.132209 -2.727026,1.738882 -0.6074272,-2.103338 -1.7775233,1.083222 0.1099487,3.008361 -6.35337885,4.307448 0.46015706,2.906545 16.80070259,-6.948921 0.882994,9.623753 -6.497939,6.438864 1.056742,1.025537 6.035765,-2.126063 c -6.9e-5,0 0.935883,1.572915 0.936578,1.57222 z"
+  },
+  $A400: {
+    x: 19,
+    y: 20,
+    scale: 0.6,
+    rotcorr: 0,
+    path: "m 10.609151,1051.8941 c 0.02312,-0.2782 0.0821,-0.6315 0.13107,-0.785 0.163705,-0.5133 0.971237,-1.2043 3.955548,-3.3847 1.549628,-1.1321 2.832287,-2.0968 2.850354,-2.1436 0.01807,-0.047 -0.09548,-0.7691 -0.252323,-1.6051 -0.801562,-4.2722 -0.912518,-5.3112 -0.913647,-8.5558 l -8.21e-4,-2.3603 -0.38728,-0.3539 c -0.512042,-0.468 -0.589696,-0.6347 -0.589696,-1.2657 v -0.5202 l -0.26169,0.029 c -0.14393,0.016 -0.889746,0.076 -1.65737,0.1347 -0.767624,0.058 -2.165049,0.1685 -3.105389,0.2444 -0.9403403,0.076 -3.5310712,0.2787 -5.757181,0.4507 -2.2261098,0.172 -4.14953142,0.3268 -4.27427033,0.3441 l -0.22679808,0.031 v -1.0457 c 0,-1.0287 0.002868,-1.0483 0.17632204,-1.2091 0.12223394,-0.1133 1.16225397,-0.4922 3.39036557,-1.2353 2.1568247,-0.7193 3.2353968,-1.1112 3.2789594,-1.1913 0.040411,-0.074 0.03885,-0.2323 -0.00414,-0.4187 -0.084642,-0.3669 -0.091233,-1.8834 -0.00964,-2.2184 0.032722,-0.1343 0.071226,-0.2908 0.085565,-0.3476 0.021026,-0.083 -0.1402442,-0.1224 -0.8332402,-0.2015 -0.4726217,-0.054 -0.8337541,-0.1234 -0.8025161,-0.1545 0.031238,-0.031 0.4257791,-0.079 0.8767582,-0.1055 l 0.8199621,-0.049 v -0.2371 c 0,-0.1558 0.056175,-0.2933 0.1636729,-0.4009 0.3101152,-0.3101 0.6737352,-0.066 0.6737352,0.4525 0,0.1992 0.01096,0.2061 0.3314745,0.2065 0.6760925,9e-4 1.4899725,0.091 1.4238515,0.1563 -0.036172,0.036 -0.4366496,0.096 -0.8899483,0.1316 -0.4532987,0.036 -0.8338619,0.075 -0.8456951,0.088 -0.011834,0.012 0.02915,0.1427 0.091073,0.291 0.091604,0.2192 0.1080329,0.4749 0.088147,1.3715 -0.020974,0.9457 -0.00959,1.1013 0.080234,1.0969 0.057573,0 0.9525514,-0.2933 1.9888437,-0.6455 l 1.884168,-0.6404 0.04411,-0.2786 c 0.02426,-0.1532 0.01175,-0.4966 -0.0278,-0.7632 -0.05857,-0.3946 -0.04908,-0.6057 0.05118,-1.138 0.06771,-0.3595 0.111615,-0.6646 0.09759,-0.6782 -0.01403,-0.014 -0.182522,-0.041 -0.374428,-0.062 -0.951561,-0.1038 -1.154518,-0.1397 -1.099702,-0.1945 0.03273,-0.033 0.395617,-0.081 0.806407,-0.1052 l 0.746894,-0.045 v -0.2337 c 0,-0.2763 0.205173,-0.5127 0.445052,-0.5127 0.212459,0 0.334985,0.1501 0.38741,0.4746 l 0.04168,0.2581 0.848855,0.069 c 0.46687,0.038 0.86584,0.088 0.886601,0.1083 0.02076,0.021 -0.343834,0.077 -0.810208,0.1232 -0.466375,0.046 -0.862941,0.1 -0.881256,0.1181 -0.01831,0.018 0.0031,0.1549 0.04766,0.3035 0.04453,0.1486 0.08097,0.6651 0.08097,1.1478 0,0.4826 0.02983,0.896 0.06631,0.9185 0.03646,0.022 0.509678,-0.1045 1.051576,-0.2823 0.858744,-0.2818 0.994798,-0.347 1.05946,-0.5079 0.171426,-0.4266 0.482992,-1.05 0.64142,-1.2835 l 0.170038,-0.2506 0.02968,-2.9192 c 0.02401,-2.3603 0.05171,-3.023 0.144746,-3.462 0.219266,-1.0345 1.118177,-2.9485 1.72438,-3.6715 0.426425,-0.5086 0.665095,-0.6071 0.91914,-0.3791 0.487951,0.4379 1.125937,1.5326 1.689772,2.8992 0.447479,1.0846 0.477771,1.3582 0.510872,4.6134 l 0.02968,2.9192 0.170037,0.2506 c 0.158428,0.2335 0.469995,0.8569 0.64142,1.2835 0.06467,0.1609 0.200716,0.2261 1.059461,0.5079 0.541898,0.1778 1.015107,0.3048 1.051575,0.2823 0.03647,-0.022 0.06631,-0.4359 0.06631,-0.9185 0,-0.4827 0.03643,-0.9992 0.08097,-1.1478 0.04453,-0.1486 0.06597,-0.2852 0.04766,-0.3035 -0.01831,-0.018 -0.41488,-0.072 -0.881255,-0.1181 -0.466374,-0.046 -0.830968,-0.1021 -0.810208,-0.1232 0.02076,-0.021 0.41973,-0.069 0.886601,-0.1083 l 0.848854,-0.069 0.04168,-0.2581 c 0.05242,-0.3245 0.174952,-0.4746 0.387411,-0.4746 0.239879,0 0.445051,0.2364 0.445051,0.5127 v 0.2337 l 0.746891,0.045 c 0.41079,0.025 0.773674,0.072 0.806407,0.1052 0.05482,0.054 -0.148141,0.091 -1.099702,0.1945 -0.191906,0.021 -0.360399,0.049 -0.374428,0.062 -0.01403,0.014 0.02988,0.3187 0.09759,0.6782 0.100255,0.5323 0.109753,0.7434 0.05118,1.138 -0.03955,0.2666 -0.05206,0.61 -0.0278,0.7632 l 0.04411,0.2786 1.884168,0.6404 c 1.036293,0.3522 1.931273,0.6427 1.988844,0.6455 0.08982,0 0.101208,-0.1512 0.08023,-1.0969 -0.01988,-0.8966 -0.0034,-1.1523 0.08815,-1.3715 0.06192,-0.1483 0.102906,-0.2792 0.09107,-0.291 -0.01183,-0.012 -0.392396,-0.051 -0.845694,-0.088 -0.453299,-0.036 -0.853776,-0.096 -0.889949,-0.1316 -0.06613,-0.066 0.747759,-0.1554 1.423853,-0.1563 0.320513,-4e-4 0.331473,-0.01 0.331473,-0.2065 0,-0.5184 0.363621,-0.7626 0.673735,-0.4525 0.107499,0.1076 0.163673,0.2451 0.163673,0.4009 v 0.2371 l 0.819962,0.049 c 0.45098,0.027 0.845521,0.075 0.876759,0.1055 0.03124,0.031 -0.329895,0.1006 -0.802516,0.1545 -0.692996,0.08 -0.854267,0.1182 -0.833241,0.2015 0.01434,0.056 0.05284,0.2133 0.08556,0.3476 0.08159,0.335 0.075,1.8515 -0.0097,2.2184 -0.04298,0.1864 -0.04454,0.3444 -0.0041,0.4187 0.04357,0.081 1.122134,0.472 3.278959,1.1913 2.228111,0.7431 3.268131,1.122 3.390365,1.2353 0.173459,0.1608 0.176323,0.1804 0.176323,1.2091 v 1.0457 l -0.226798,-0.031 c -0.12474,-0.017 -2.048161,-0.1721 -4.274271,-0.3441 -2.22611,-0.172 -4.816841,-0.3747 -5.757181,-0.4507 -0.940339,-0.076 -2.337764,-0.1859 -3.105388,-0.2444 -0.767624,-0.058 -1.513441,-0.119 -1.65737,-0.1347 l -0.26169,-0.029 v 0.5202 c 0,0.631 -0.07766,0.7977 -0.589696,1.2657 l -0.387283,0.3539 -8.24e-4,2.3603 c -0.0011,3.2446 -0.112085,4.2836 -0.913647,8.5558 -0.156844,0.836 -0.270389,1.5583 -0.252322,1.6051 0.01807,0.046 1.300725,1.0115 2.850353,2.1436 2.984311,2.1804 3.791844,2.8714 3.955547,3.3847 0.120421,0.3775 0.193545,1.291 0.103345,1.291 -0.06512,0 -1.762262,-0.6231 -6.004366,-2.2043 -0.969246,-0.3613 -1.790511,-0.6569 -1.825032,-0.6569 -0.03452,0 -0.129532,0.2198 -0.211138,0.4885 -0.08161,0.2687 -0.177969,0.4884 -0.214141,0.4884 -0.03617,0 -0.132535,-0.2197 -0.21414,-0.4884 -0.08161,-0.2687 -0.176619,-0.4885 -0.211139,-0.4885 -0.03452,0 -0.855785,0.2956 -1.825033,0.6569 -4.242104,1.5812 -5.939249,2.2043 -6.004368,2.2043 -0.04394,0 -0.0542,-0.1872 -0.02772,-0.506 z"
+  },
+  $B737: {
+    x: 20,
+    y: 20,
+    scale: 0.76,
+    rotcorr: 0,
+    path: "m 32,4.8 0.3,0.1 0.8,1.2 0.9,3.1 0.6,3.1 0.5,5.2 0,6.9 3.6,2.7 -0.2,-1.5 0,-1.5 0.1,-1.2 0.2,-0.4 2.8,0 0.2,0.4 0.1,1.2 0,1.5 -0.4,2.2 -0.3,0 -0.1,0.9 14,7.7 0.4,0.4 0.2,0.7 0,1.5 -8.3,-2.4 -0.1,0.4 -0.2,0.2 -0.2,-0.2 -0.1,-0.5 -3.7,-1.1 -0.1,0.4 -0.2,0.2 -0.2,-0.2 -0.1,-0.5 -2,-0.6 0,1 -0.1,0.5 -0.1,0.1 -0.4,0 -0.1,-0.1 -0.1,-0.5 0,-1.3 -4.6,0 0,7.7 -0.1,2.8 -0.6,3.7 -0.6,3 0.1,0.4 8.5,5.9 0,1.8 -10,-2.5 -0.2,0.7 -0.1,0.8 -0.1,0.4 -0.1,-0.4 -0.1,-0.8 -0.2,-0.7 -10,2.5 0,-1.8 8.5,-5.9 0.1,-0.4 -0.6,-3 -0.6,-3.7 -0.1,-2.8 0,-7.7 -4.6,0 0,1.3 -0.1,0.5 -0.1,0.1 -0.4,0 -0.1,-0.1 -0.1,-0.5 0,-1 -2,0.6 -0.1,0.5 -0.2,0.2 -0.2,-0.2 -0.1,-0.4 -3.7,1.1 -0.1,0.5 -0.2,0.2 -0.2,-0.2 -0.1,-0.4 -8.3,2.4 0,-1.5 0.2,-0.7 0.4,-0.4 14,-7.7 -0.1,-0.9 -0.3,0 -0.4,-2.2 0,-1.5 0.1,-1.2 0.2,-0.4 2.8,0 0.2,0.4 0.1,1.6 0,1.1 -0.2,1.5 3.6,-2.8 0,-6.8 0.5,-5.2 0.6,-3.1 1,-3.1 0.7,-1.2 z"
+  },
+  $B747: {
+    x: 20,
+    y: 20,
+    scale: 0.88,
+    rotcorr: 0,
+    path: "m 20.170952,35.711605 c 0.240955,-0.452026 0.835455,-1.57222 0.835455,-1.57222 l 6.03507,2.126063 1.056047,-1.025537 -6.496618,-6.438864 0.883688,-9.623753 16.799383,6.948921 0.461547,-2.906545 -6.354074,-4.307448 0.109254,-3.008361 -1.776828,-1.083222 -0.608121,2.103338 -2.727027,-1.738882 0.337976,-3.132209 -1.375745,-1.158211 -1.689327,2.551609 -3.321597,-3.620931 0.01293,-3.9857346 -1.062159,-4.86699002 -1.125964,-0.70583829 h -0.08889 l -1.125965,0.70583829 -1.060843,4.86699002 0.01153,3.9857346 -3.322291,3.620931 -1.687937,-2.551609 -1.37505,1.158211 0.335961,3.132209 -2.727026,1.738882 -0.6074272,-2.103338 -1.7775233,1.083222 0.1099487,3.008361 -6.35337885,4.307448 0.46015706,2.906545 16.80070259,-6.948921 0.882994,9.623753 -6.497939,6.438864 1.056742,1.025537 6.035765,-2.126063 c -6.9e-5,0 0.935883,1.572915 0.936578,1.57222 z"
+  },
+  $B777: {
+    x: 20,
+    y: 20,
+    scale: 0.7,
+    rotcorr: 0,
+    path: "m 32,0.2 0.1,0 0.2,0.1 0.3,0.3 0.5,0.8 0.7,1.7 0.5,2 0.3,2.5 0,15 0.3,0.6 0.5,0.6 3.3,2.2 -0.2,-1.1 0,-1.6 0.2,-1.4 0.1,-0.2 3,0 0.1,0.2 0.2,1.4 0,1.6 -0.2,1.4 -0.4,0 -0.4,1.3 16.4,10.8 0.2,0.5 0.1,1.5 -9,-3.1 -0.1,0.6 -0.1,0.2 -0.1,-0.2 -0.1,-0.7 -4.2,-1.4 -0.1,0.6 -0.1,0.2 -0.1,-0.2 -0.1,-0.7 -2.5,-0.8 -1.8,0 -0.1,0.7 -0.1,0.2 -0.1,-0.2 -0.1,-0.7 -4.5,0 0,14.9 -0.2,1.9 -0.7,3.8 7.6,6.2 0.1,0.7 0,0.7 -0.1,0.7 -8.5,-3.1 -0.2,0.4 -0.5,1.8 -0.2,0 -0.5,-1.8 -0.2,-0.4 -8.5,3.1 -0.1,-0.7 0,-0.7 0.1,-0.7 7.6,-6.2 -0.7,-3.8 -0.2,-1.9 0,-14.9 -4.5,0 -0.1,0.7 -0.1,0.2 -0.1,-0.2 -0.1,-0.7 -1.8,0 -2.5,0.8 -0.1,0.7 -0.1,0.2 -0.1,-0.2 -0.1,-0.6 -4.2,1.4 -0.1,0.7 -0.1,0.2 -0.1,-0.2 -0.1,-0.6 -9,3.1 0.1,-1.5 0.2,-0.5 16.4,-10.8 -0.4,-1.3 -0.4,0 -0.2,-1.4 0,-1.6 0.2,-1.4 0.1,-0.2 3,0 0.1,0.2 0.2,1.4 0,1.6 -0.2,1.1 3.3,-2.2 0.5,-0.6 0.3,-0.6 0,-15 0.3,-2.5 0.5,-2 0.7,-1.7 0.5,-0.8 0.3,-0.3 0.2,-0.1 z"
+  },
+  $B787: {
+    x: 20,
+    y: 20,
+    scale: 0.7,
+    rotcorr: 0,
+    path: "m 32,0.2 0.1,0 0.4,0.3 0.7,1 0.8,1.9 0.6,2.5 0.2,2.5 0.1,2.7 0,11.4 6.5,4.6 -0.3,-0.8 -0.2,0.3 -0.2,-0.3 -0.2,0.1 -0.1,-0.3 -0.2,-0.9 -0.1,-1.3 0,-1.6 0.1,-1.2 0.1,-0.4 0.2,-0.2 3.2,0 0.2,0.2 0.1,0.4 0.2,1.2 0,1.6 -0.1,1.3 -0.4,1.4 -0.1,-0.3 -0.1,0.3 -0.1,-0.3 -0.2,0.3 -0.2,-0.3 -0.4,1.6 17.5,12.2 0.6,0.6 0.4,0.6 1.2,2 0,1 -1.6,-1.4 -0.7,-0.5 -0.9,-0.5 -7.5,-3.2 -0.1,0.8 -0.2,0.3 -0.3,-0.6 -0.1,-0.8 -3.8,-1.6 -0.1,0.8 -0.2,0.3 -0.3,-0.7 -0.1,-0.7 -3.2,-1.4 -2,0 -0.1,0.7 -0.2,0.4 -0.2,-0.4 -0.1,-0.7 -5.7,0 0,13 -0.1,1.8 -0.1,1.6 -0.5,3.3 0,0.4 0.2,0.4 7,6.4 0.2,0.3 0.1,0.3 0.4,1.3 -9.2,-3.4 -0.7,2.9 -0.4,0 -0.7,-2.9 -9.2,3.4 0.4,-1.3 0.1,-0.3 0.2,-0.3 7,-6.4 0.2,-0.4 0,-0.4 -0.5,-3.3 -0.1,-1.7 -0.1,-1.7 0,-13 -5.7,0 -0.1,0.7 -0.2,0.5 -0.2,-0.5 -0.1,-0.7 -2,0 -3.3,1.4 -0.1,0.6 -0.3,0.6 -0.2,-0.3 -0.1,-0.7 -3.6,1.7 -0.1,0.6 -0.3,0.6 -0.2,-0.3 -0.1,-0.6 -7.6,3.2 -0.9,0.5 -0.7,0.5 -1.6,1.4 0,-1 1.2,-2 0.4,-0.6 0.6,-0.6 17.5,-12.2 -0.4,-1.6 -0.2,0.3 -0.2,-0.3 -0.1,0.3 -0.1,-0.3 -0.1,0.3 -0.3,-1.4 -0.1,-1.3 0,-1.6 0.1,-1.2 0.1,-0.4 0.2,-0.2 3.2,0 0.2,0.2 0.1,0.4 0.1,1.2 0,1.6 -0.1,1.3 -0.2,0.9 -0.1,0.3 -0.2,-0.1 -0.2,0.3 -0.2,-0.3 -0.3,0.8 6.5,-4.6 0,-11.4 0.1,-2.7 0.2,-2.5 0.6,-2.5 0.8,-1.9 0.7,-1 0.4,-0.3 z"
+  },
+  $F15: {
+    x: 20,
+    y: 20,
+    scale: 0.6,
+    rotcorr: 0,
+    path: "m 12.720751,1092.038 c -0.356706,-0.4854 -0.956831,-1.7747 -0.956831,-2.0557 0,-0.1857 0.720533,-1.1226 1.769853,-2.3012 1.376792,-1.5465 1.790412,-1.9294 1.862425,-1.7241 0.08054,0.2296 0.135532,0.2093 0.423233,-0.1565 0.241221,-0.3067 0.33122,-0.6296 0.332727,-1.194 0.0011,-0.4255 0.04311,-0.9268 0.09326,-1.114 0.09084,-0.339 0.08419,-0.3403 -1.703598,-0.3387 -1.58712,0 -2.01243,0.063 -3.675647,0.5294 -1.0344685,0.2903 -1.937655,0.5278 -2.0070814,0.5278 -0.069427,0 -0.4651626,-0.5192 -0.8794134,-1.1538 -1.2438223,-1.9054 -1.4867824,-1.4639 3.7086138,-6.7391 l 4.461797,-4.5303 0.0032,-1.3886 c 0.0017,-0.7637 0.105507,-1.831 0.230609,-2.3717 0.125101,-0.5407 0.287534,-1.5392 0.360962,-2.2189 0.200087,-1.8521 0.207282,-1.8648 1.056305,-1.8648 l 0.723159,0 0.08939,-2.609 c 0.08714,-2.5434 0.382385,-4.722 0.978724,-7.2221 0.377706,-1.5834 0.616407,-1.5151 1.056703,0.3025 0.559043,2.3079 0.796043,4.1347 0.897703,6.9196 l 0.09524,2.609 0.723792,0 c 0.398086,0 0.763755,0.065 0.812599,0.1437 0.04884,0.079 0.15182,0.7426 0.228839,1.4747 0.07702,0.732 0.242864,1.7733 0.368547,2.314 0.125685,0.5407 0.230908,1.6421 0.23383,2.4474 l 0.0053,1.4642 4.487005,4.5559 c 3.723507,3.7807 4.487004,4.6278 4.487004,4.9784 0,0.3836 -1.249116,2.6193 -1.576409,2.8216 -0.07257,0.045 -1.00769,-0.1599 -2.078051,-0.455 -1.690712,-0.4661 -2.167461,-0.5365 -3.632827,-0.5365 l -1.686718,0 0,1.1836 c 0,1.0551 0.03957,1.2208 0.364428,1.526 0.326381,0.3066 0.369806,0.3148 0.41593,0.078 0.07357,-0.3771 3.607343,3.609 3.601563,4.0627 -0.0075,0.5848 -0.940546,2.1432 -1.315161,2.1964 -0.184968,0.026 -1.099991,-0.2285 -2.033383,-0.5664 -1.005037,-0.3637 -1.73431,-0.554 -1.788381,-0.4665 -0.05021,0.081 -0.291407,-0.4295 -0.535983,-1.1351 l -0.444682,-1.2828 -1.098381,0 c -0.776638,0 -1.099059,-0.055 -1.100697,-0.1891 -0.0018,-0.1442 -0.02961,-0.1461 -0.117357,-0.01 -0.06425,0.1011 -0.547372,0.2013 -1.094228,0.2269 l -0.979186,0.046 -0.437272,1.26 c -0.2405,0.693 -0.481095,1.1891 -0.534655,1.1024 -0.06232,-0.1008 -0.691537,0.058 -1.747806,0.4416 -1.750261,0.6353 -2.224968,0.7148 -2.448953,0.41 z"
+  },
+  $F16: {
+    x: 20,
+    y: 20,
+    scale: 0.6,
+    rotcorr: 0,
+    path: "m 19.808754,1092.7558 c -1.116025,-0.7717 -1.099627,-2.6125 -1.388116,-4.1313 -0.216638,0.15 -0.08738,1.7345 -0.121449,2.4897 0.06919,0.7244 -1.556603,-0.017 -1.097866,1.1613 -1.389019,-0.041 -2.798968,0.1089 -4.17032,-0.1332 -0.974564,-0.2458 -0.988329,-1.8214 -0.53518,-2.4642 1.488914,-1.3497 3.118483,-2.5579 4.488518,-4.0271 0.441692,-1.0582 0.12943,-2.2649 0.21696,-3.3935 -3.198416,0.028 -6.400993,-0.068 -9.5961409,0.071 -1.280678,1.3674 -0.7004539,-1.7519 -0.8334432,-2.5906 0.1178249,-1.2981 -0.2295071,-3.4192 0.2687784,-4.2446 0.3169018,1.1184 0.225064,2.3043 0.3176875,3.4564 2.9753972,-2.3363 5.9022912,-4.7335 8.8488602,-7.1059 0.532037,-3.1529 1.57549,-6.2104 1.845912,-9.4129 0.375408,-3.2796 0.910155,-6.5512 1.94412,-9.6937 1.343724,3.6518 1.820013,7.5296 2.265927,11.3726 0.326638,2.597 1.227203,5.0913 1.632855,7.6615 2.949327,2.3913 5.854498,4.8624 8.859904,7.1922 0.187314,-0.7479 -0.0037,-3.6004 0.445602,-3.149 0.06075,2.284 0.132367,4.5839 -0.0424,6.862 -0.728995,-0.5735 -2.738544,-0.395 -4.011314,-0.4191 -2.060852,-7e-4 -4.121705,-10e-5 -6.182557,-3e-4 0,1.1348 0,2.2696 0,3.4044 1.623901,1.4078 3.247827,2.8156 4.87178,4.2233 0.235392,1.1743 -0.04572,2.6942 -1.637626,2.3754 -1.123707,0.017 -2.247596,0.01 -3.371392,0.01 0.478177,-1.2766 -1.20092,-0.4087 -1.10113,-1.2225 -0.007,-0.8506 -0.01394,-1.7012 -0.02092,-2.5518 -0.274372,1.2026 -0.346042,2.482 -0.681696,3.6392 -1.00129,-0.4434 -0.822888,2.0076 -1.215389,0.615 l 3e-5,0.01 z"
+  },
+  $TORN: {
+    x: 20,
+    y: 20,
+    scale: 0.6,
+    rotcorr: 0,
+    path: "m 19.995751,1092.1652 c -0.150693,-1.6016 -0.08058,-3.2198 -0.1722,-4.8281 -1.435177,-0.1548 -2.334768,0.096 -3.593659,0.738 -1.332791,0.2327 -3.388896,1.5828 -4.254408,1.0149 0.06317,-0.8387 -0.330774,-1.8055 0.319992,-2.4951 1.445853,-2.4474 3.173451,-4.7106 4.818235,-7.0248 -0.105269,-0.8498 0.265646,-2.2343 -0.274174,-2.6551 -4.35855,1.0889 -8.6947595,2.265 -13.0469343,3.3791 0.2752673,-1.3096 0.1255747,-3.1776 1.4276044,-3.9076 3.1313154,-1.5014 6.3141929,-2.8943 9.4348289,-4.418 0.614811,-0.8539 1.25511,-1.7693 1.688462,-2.7128 0,-1.3831 0,-2.7663 0,-4.1494 0.562003,-0.076 1.412562,0.1466 1.803279,-0.1059 0.1352,-3.0038 0.118447,-6.077 0.957159,-8.9826 0.617533,-0.9515 0.789054,-2.9963 1.070414,-3.5253 -0.04074,2.0653 1.400169,3.8144 1.448272,5.8878 0.297231,2.2286 0.422099,4.4782 0.392472,6.726 0.56773,0.074 1.42962,-0.1453 1.820515,0.106 0.03689,1.3932 -0.07358,2.8278 0.0551,4.1953 0.572757,0.8578 1.005564,1.8574 1.711527,2.5921 3.292599,1.6264 6.724724,2.9879 9.942281,4.7532 0.473575,1.1145 1.798717,4.4977 -0.657184,3.1357 -3.915085,-0.9914 -7.811706,-2.0571 -11.734415,-3.0171 -0.103419,0.8746 -0.152392,1.8781 -0.02561,2.7464 1.770504,2.4677 3.594992,4.9053 5.11955,7.5368 -0.09209,0.9856 0.456922,3.0211 -1.134027,1.9715 -1.517788,-0.6641 -3.50429,-0.8187 -4.705213,-1.8626 -0.70794,0.3244 -2.325963,-0.4716 -2.008033,0.8149 -0.104053,1.3913 0.04844,2.8071 -0.181626,4.1819 l -0.0934,0.083 -0.1288,-0.1783 -1.9e-5,-2e-4 z"
+  },
+  $EUFI: {
+    x: 20,
+    y: 20,
+    scale: 0.6,
+    rotcorr: 0,
+    path: "m 20.049625,0.19319282 c 0.346836,1.81282578 0.951463,3.58610348 1.091119,5.43071318 1.060626,1.1387891 2.380342,2.1611588 3.2809,3.3718727 0.291748,1.2113483 -0.04932,1.5817703 -1.164136,0.9239813 -0.748347,-0.1581792 -2.137578,-1.1703168 -1.684016,0.273009 -0.0054,0.905054 0.971399,1.118925 0.614463,2.246232 0.02479,0.899759 -0.128028,1.861658 0.656311,2.495684 3.225087,4.268534 6.450174,8.537069 9.675261,12.805603 0.228397,-1.106513 0.462564,-1.848707 0.732895,-0.307602 0.05485,1.357252 0.126351,2.778961 -0.03812,4.127018 -0.335386,1.737401 -0.285809,-0.857542 -1.273962,-0.317361 -2.92441,0.196885 -5.84882,0.393769 -8.77323,0.590654 0.461145,1.176313 -0.443713,1.255733 -0.594648,2.073835 -0.283981,0.641309 0.308782,2.236929 -0.62009,2.17629 -1.101382,-0.242451 -1.654134,0.137781 -1.521668,1.29572 -0.109454,0.686238 -0.2329,2.620688 -0.315462,0.907759 -0.287601,-0.676451 0.287171,-2.351265 -0.801459,-2.10037 -1.200269,0.364073 -1.519762,-0.271966 -1.413234,-1.391748 0.302623,-1.11103 -0.684152,-1.367732 -0.766458,-2.203109 0.324271,-1.285981 -1.141784,-0.758377 -1.931002,-0.916403 C 12.713933,31.570233 10.224777,31.465495 7.7356214,31.360757 7.5643096,32.807462 7.2095659,32.413311 7.03419,31.215633 7.0740767,29.604579 6.8725823,27.968632 7.3459376,26.40228 c 0.2070037,0.874599 0.3208355,2.212353 0.9961255,0.75912 3.1751149,-4.267045 6.3502299,-8.534091 9.5253449,-12.801136 0.104216,-1.109643 -0.09944,-2.381426 0.195639,-3.390552 0.772772,-0.336901 0.536273,-2.1860426 -0.482118,-1.3111832 -0.646351,-0.024115 -2.166007,1.2151062 -2.00955,0.029634 -0.282417,-1.0363227 0.99479,-1.4765273 1.507298,-2.2090146 0.585866,-0.6838264 1.468619,-1.2435317 1.879841,-2.0003726 0.02183,-1.788216 0.764573,-3.5298844 1.091107,-5.28558278 z"
+  },
+  $BEACON: {
+    x: 20,
+    y: 20,
+    scale: 0.1,
+    rotcorr: 90,
+    path: "M 177,126 L 172,197 L 164,197 L 164,203 L 188,203 C 189,208 194,213 200,213 C 206,217 211,208 212,203 L 236,203 L 236,197 L 228,197 L 223,126 L 177,126 z M 183,133 L 217,133 L 221,197 L 212,197 C 211,191 206,187 200,187 C 194,188 189,192 188,197 L 179,197 L 183,133 z M 200,193 C 204,193 208,196 208,200 C 208,204 204,208 200,208 C 196,208 193,204 193,200 C 193,196 196,193 200,193 z"
+  },
+  $CAR: {
+    x: 0,
+    y: 0,
+    scale: 2,
+    rotcorr: -180,
+    path: "M0,1 2,1 2,0 4,0 4,1 8,1 8,0 10,0 10,1 12,1 12,3 9,3 6,5                 2,5 1,3 0,3z"
+  },
+  // TRUCK from https://icomoon.io/app/#/select
+  $TRUCK: {
+    x: 12,
+    y: 12,
+    scale: 0.1,
+    rotcorr: -90,
+    path: "M315.075 263.385l-14.769-29.538h-22.153v-14.768c0-4.061-3.324-7.385-7.385-7.385h-66.459c-4.061 0-7.385 3.323-7.385 7.385v59.075l7.385 7.385h9.365c-1.258 2.173-1.981 4.693-1.981 7.385 0 8.156 6.613 14.769 14.768 14.769s14.768-6.613 14.768-14.769c0-2.689-0.725-5.212-1.981-7.385h40.883c-1.258 2.173-1.981 4.693-1.981 7.385 0 8.156 6.613 14.769 14.769 14.769s14.769-6.613 14.769-14.769c0-2.689-0.724-5.212-1.981-7.385h9.365v-22.153zM278.153 263.385v-22.153h15.307l11.075 22.153h-26.384z"
+  },
+  $ALARMCAR: {
+    x: 0,
+    y: 0,
+    scale: 2,
+    rotcorr: -90,
+    path: "M0,1 2,1 2,0 4,0 4,1 8,1 8,0 10,0 10,1 12,1 12,3 9,3 6,5 6,7 4,7 4,5 3,5 2,5 1,3 0,3z"
+  },
+  $PARAGLIDER: {
+    x: 0,
+    y: 0,
+    scale: 0.01,
+    rotcorr: -90,
+    path: "M256.001,0C142.514,0,50.186,92.328,50.186,205.814c0,1.456,0.348,2.892,1.014,4.186l0.798,1.551 c0.154,0.3,0.325,0.591,0.511,0.871l195.784,295.246c1.009,1.62,2.487,2.87,4.227,3.598c0.121,0.051,0.239,0.105,0.362,0.15 c0.233,0.085,0.469,0.16,0.708,0.227c0.195,0.055,0.393,0.101,0.593,0.143c0.201,0.042,0.4,0.085,0.605,0.115 c0.369,0.052,0.743,0.082,1.12,0.089c0.045,0,0.088,0.011,0.132,0.011c0.005,0,0.011-0.001,0.016-0.001 c0.006,0,0.013,0.001,0.02,0.001c0.001,0,0.002,0,0.004,0c0.454,0,0.9-0.044,1.342-0.11c0.071-0.011,0.14-0.027,0.211-0.04 c0.395-0.068,0.782-0.162,1.162-0.281c0.051-0.016,0.102-0.032,0.154-0.049c1.922-0.637,3.612-1.888,4.758-3.621l196.589-297.036 c0.992-1.498,1.52-3.254,1.52-5.05C461.815,92.328,369.487,0,256.001,0z M223.634,31.592c7.533-6.373,15.31-10.463,23.214-12.248 v167.661c-11.956-7.188-25.667-11.035-40.013-11.036c-14.159,0-27.701,3.751-39.547,10.758 c2.145-43.896,11.669-84.492,27.365-115.884C203.126,53.897,212.877,40.692,223.634,31.592z M203.086,25.917 c-9.085,9.706-17.437,22.011-24.802,36.742c-16.977,33.953-27.201,77.647-29.339,124.599 c-12.055-7.354-25.924-11.289-40.442-11.289c-13.945,0-27.288,3.643-39.004,10.448C77.388,109.991,131.362,47.049,203.086,25.917z M71.159,207.431c10.547-8.521,23.563-13.16,37.344-13.16c15.119,0,29.327,5.569,40.352,15.74l-0.804,0.27l77.087,229.355 L71.159,207.431z M246.85,446.784l-79.761-237.309c10.939-9.825,24.899-15.204,39.746-15.204c14.966,0,29.037,5.464,40.014,15.442 V446.784z M265.15,19.344c7.905,1.785,15.682,5.875,23.216,12.248c10.757,9.1,20.508,22.305,28.981,39.251 c15.696,31.393,25.225,71.99,27.369,115.886c-11.845-7.008-25.389-10.759-39.55-10.76c-14.346,0-28.06,3.848-40.016,11.036V19.344 z M265.15,209.714c10.978-9.978,25.047-15.442,40.016-15.442c14.847,0,28.809,5.381,39.747,15.205L265.15,446.974V209.714z M286.903,439.69l77.046-229.408l-0.804-0.27c11.025-10.172,25.234-15.74,40.352-15.74c13.718,0,26.678,4.597,37.201,13.044 L286.903,439.69z M403.497,175.969c-14.516,0-28.387,3.935-40.442,11.289c-2.138-46.952-12.362-90.647-29.339-124.599 c-7.366-14.733-15.721-27.036-24.809-36.743c71.728,21.13,125.705,84.075,133.596,160.502 C430.786,179.612,417.442,175.969,403.497,175.969z"
+  }
+};
+
+var ac_symbol_map = [
+  { id: "A109", sym: "$7" },
+  { id: "A119", sym: "$7" },
+  { id: "A124", sym: "$5" },
+  { id: "A129", sym: "$7" },
+  { id: "A139", sym: "$7" },
+  { id: "A149", sym: "$7" },
+  { id: "A169", sym: "$7" },
+  { id: "A2RT", sym: "$7" },
+  { id: "A30", sym: "$4" },
+  { id: "A31", sym: "$A320" },
+  { id: "A32", sym: "$A320" },
+  { id: "A33", sym: "$A330" },
+  { id: "A35", sym: "$A330" },
+  { id: "A34", sym: "$A340" },
+  { id: "A38", sym: "$B747" },
+  { id: "A400", sym: "$A400" },
+  { id: "A3ST", sym: "$A330" },
+  { id: "AA5", sym: "$1" },
+  { id: "AC11", sym: "$1" },
+  { id: "AJET", sym: "$6" },
+  { id: "ALH", sym: "$7" },
+  { id: "ALO2", sym: "$7" },
+  { id: "ALO3", sym: "$7" },
+  { id: "AN12", sym: "$C" },
+  { id: "AN24", sym: "$B" },
+  { id: "AN26", sym: "$9" },
+  { id: "ANST", sym: "$7" },
+  { id: "AS32", sym: "$7" },
+  { id: "AS3B", sym: "$7" },
+  { id: "AS50", sym: "$7" },
+  { id: "AS55", sym: "$7" },
+  { id: "AS65", sym: "$7" },
+  { id: "AT4", sym: "$D" },
+  { id: "AT7", sym: "$D" },
+  { id: "ATP", sym: "$D" },
+  { id: "B06", sym: "$7" },
+  { id: "B06T", sym: "$7" },
+  { id: "B105", sym: "$7" },
+  { id: "B190", sym: "$2" },
+  { id: "B212", sym: "$7" },
+  { id: "B214", sym: "$7" },
+  { id: "B222", sym: "$7" },
+  { id: "B230", sym: "$7" },
+  { id: "B305", sym: "$7" },
+  { id: "B350", sym: "$2" },
+  { id: "B407", sym: "$7" },
+  { id: "B412", sym: "$7" },
+  { id: "B427", sym: "$7" },
+  { id: "B429", sym: "$7" },
+  { id: "B430", sym: "$7" },
+  { id: "B461", sym: "$9" },
+  { id: "B462", sym: "$9" },
+  { id: "B463", sym: "$9" },
+  { id: "B47G", sym: "$7" },
+  { id: "B47J", sym: "$7" },
+  { id: "B47T", sym: "$7" },
+  { id: "B70", sym: "$5" },
+  { id: "B712", sym: "$3" },
+  { id: "B72", sym: "$3" },
+  { id: "B720", sym: "$5" },
+  { id: "B73", sym: "$B737" },
+  { id: "B74", sym: "$B747" },
+  { id: "B75", sym: "$4" },
+  { id: "B76", sym: "$4" },
+  { id: "B77", sym: "$B777" },
+  { id: "B78", sym: "$B787" },
+  { id: "BA11", sym: "$3" },
+  { id: "BABY", sym: "$7" },
+  { id: "BDOG", sym: "$1" },
+  { id: "BE19", sym: "$2" },
+  { id: "BE20", sym: "$2" },
+  { id: "BE33", sym: "$1" },
+  { id: "BE36", sym: "$1" },
+  { id: "BE40", sym: "$8" },
+  { id: "BE58", sym: "$2" },
+  { id: "BE9", sym: "$2" },
+  { id: "BK17", sym: "$7" },
+  { id: "BLCF", sym: "$5" },
+  { id: "BN2P", sym: "$D" },
+  { id: "BN2T", sym: "$D" },
+  { id: "BRB2", sym: "$7" },
+  { id: "BREZ", sym: "$1" },
+  { id: "BSTP", sym: "$7" },
+  { id: "C120", sym: "$1" },
+  { id: "C130", sym: "$2" },
+  { id: "C140", sym: "$1" },
+  { id: "C150", sym: "$1" },
+  { id: "C152", sym: "$1" },
+  { id: "C160", sym: "$B" },
+  { id: "C17", sym: "$5" },
+  { id: "C172", sym: "$1" },
+  { id: "C180", sym: "$1" },
+  { id: "C182", sym: "$1" },
+  { id: "C190", sym: "$1" },
+  { id: "C208", sym: "$1" },
+  { id: "C25A", sym: "$8" },
+  { id: "C25B", sym: "$8" },
+  { id: "C25C", sym: "$8" },
+  { id: "C303", sym: "$2" },
+  { id: "C30J", sym: "$C" },
+  { id: "C310", sym: "$2" },
+  { id: "C402", sym: "$2" },
+  { id: "C404", sym: "$2" },
+  { id: "C42", sym: "$1" },
+  { id: "C421", sym: "$2" },
+  { id: "C425", sym: "$2" },
+  { id: "C5", sym: "$5" },
+  { id: "C55B", sym: "$3" },
+  { id: "C500", sym: "$8" },
+  { id: "C501", sym: "$8" },
+  { id: "C51", sym: "$8" },
+  { id: "C525", sym: "$8" },
+  { id: "C550", sym: "$8" },
+  { id: "C560", sym: "$8" },
+  { id: "C56X", sym: "$8" },
+  { id: "C650", sym: "$8" },
+  { id: "C680", sym: "$8" },
+  { id: "C72", sym: "$1" },
+  { id: "C750", sym: "$8" },
+  { id: "C77R", sym: "$1" },
+  { id: "C90", sym: "$2" },
+  { id: "CH7", sym: "$7" },
+  { id: "CHIF", sym: "$7" },
+  { id: "CL30", sym: "$8" },
+  { id: "CL60", sym: "$8" },
+  { id: "CL61", sym: "$8" },
+  { id: "CL64", sym: "$8" },
+  { id: "CL65", sym: "$8" },
+  { id: "CN35", sym: "$B" },
+  { id: "COMU", sym: "$7" },
+  { id: "CRJ1", sym: "$3" },
+  { id: "CRJ2", sym: "$3" },
+  { id: "CRJ7", sym: "$3" },
+  { id: "CRJ9", sym: "$3" },
+  { id: "CRUZ", sym: "$1" },
+  { id: "D11", sym: "$1" },
+  { id: "D228", sym: "$D" },
+  { id: "D28", sym: "$D" },
+  { id: "D328", sym: "$D" },
+  { id: "DA40", sym: "$1" },
+  { id: "DA42", sym: "$2" },
+  { id: "DC10", sym: "$4" },
+  { id: "DC8", sym: "$5" },
+  { id: "DC9", sym: "$3" },
+  { id: "DH8A", sym: "$D" },
+  { id: "DH8B", sym: "$D" },
+  { id: "DH8C", sym: "$D" },
+  { id: "DH8D", sym: "$D" },
+  { id: "DHC1", sym: "$1" },
+  { id: "DHC6", sym: "$2" },
+  { id: "DHC7", sym: "$D" },
+  { id: "DHC8", sym: "$D" },
+  { id: "DJIN", sym: "$7" },
+  { id: "DOVE", sym: "$2" },
+  { id: "DR10", sym: "$1" },
+  { id: "DR22", sym: "$1" },
+  { id: "DR30", sym: "$1" },
+  { id: "DR40", sym: "$1" },
+  { id: "DRAG", sym: "$7" },
+  { id: "DYH2", sym: "$7" },
+  { id: "E110", sym: "$2" },
+  { id: "E120", sym: "$2" },
+  { id: "E121", sym: "$2" },
+  { id: "E135", sym: "$3" },
+  { id: "E145", sym: "$3" },
+  { id: "E170", sym: "$4" },
+  { id: "E175", sym: "$4" },
+  { id: "E190", sym: "$4" },
+  { id: "E195", sym: "$4" },
+  { id: "E300", sym: "$1" },
+  { id: "E35", sym: "$3" },
+  // { id: "E3CF", sym: "$AWACS" },
+  // { id: "E3TF", sym: "$AWACS" },
+  { id: "E400", sym: "$1" },
+  { id: "E50P", sym: "$8" },
+  { id: "E55P", sym: "$8" },
+  // { id: "E6", sym: "$AWACS" },
+  { id: "E75S", sym: "$4" },
+  { id: "EA50", sym: "$8" },
+  { id: "EA55", sym: "$8" },
+  { id: "EC20", sym: "$7" },
+  { id: "EC25", sym: "$7" },
+  { id: "EC30", sym: "$7" },
+  { id: "EC35", sym: "$7" },
+  { id: "EC45", sym: "$7" },
+  { id: "EC55", sym: "$7" },
+  { id: "EH10", sym: "$E" },
+  { id: "ELTO", sym: "$7" },
+  { id: "EN28", sym: "$7" },
+  { id: "EN48", sym: "$7" },
+  { id: "ES11", sym: "$7" },
+  // { id: "EUFI", sym: "$EUFI" },
+  { id: "EUPA", sym: "$1" },
+  { id: "EXEC", sym: "$7" },
+  { id: "EXEJ", sym: "$7" },
+  { id: "EXPL", sym: "$7" },
+  { id: "F100", sym: "$3" },
+  // { id: "F15", sym: "$F15" },
+  // { id: "F15E", sym: "$F15" },
+  // { id: "F16", sym: "$F16" },
+  { id: "F27", sym: "$D" },
+  { id: "F2EX", sym: "$8" },
+  { id: "F2TH", sym: "$8" },
+  { id: "F406", sym: "$2" },
+  { id: "F50", sym: "$D" },
+  { id: "F70", sym: "$3" },
+  { id: "F900", sym: "$8" },
+  { id: "FA10", sym: "$8" },
+  { id: "FA20", sym: "$8" },
+  { id: "FA50", sym: "$8" },
+  { id: "FA7", sym: "$8" },
+  { id: "FDCT", sym: "$1" },
+  { id: "FH11", sym: "$7" },
+  { id: "FREL", sym: "$7" },
+  { id: "G109", sym: "$1" },
+  { id: "G115", sym: "$1" },
+  { id: "G120", sym: "$1" },
+  { id: "G150", sym: "$8" },
+  { id: "G200", sym: "$8" },
+  { id: "G280", sym: "$8" },
+  { id: "G2CA", sym: "$7" },
+  { id: "G450", sym: "$8" },
+  { id: "G550", sym: "$8" },
+  { id: "G650", sym: "$8" },
+  { id: "GA8", sym: "$1" },
+  { id: "GRND", sym: "$0" },
+  { id: "GALX", sym: "$8" },
+  { id: "GAZL", sym: "$7" },
+  { id: "GL5T", sym: "$8" },
+  { id: "GLEX", sym: "$8" },
+  { id: "GLF", sym: "$8" },
+  { id: "GZLE", sym: "$7" },
+  { id: "H12T", sym: "$7" },
+  { id: "H160", sym: "$7" },
+  { id: "H2", sym: "$7" },
+  { id: "H21", sym: "$7" },
+  { id: "H25", sym: "$8" },
+  { id: "H269", sym: "$7" },
+  { id: "H43A", sym: "$7" },
+  { id: "H43B", sym: "$7" },
+  { id: "H46", sym: "$E" },
+  { id: "H47", sym: "$E" },
+  { id: "H500", sym: "$7" },
+  { id: "H53", sym: "$E" },
+  { id: "H60", sym: "$E" },
+  { id: "H64", sym: "$7" },
+  { id: "H750", sym: "$8" },
+  { id: "H900", sym: "$3" },
+  { id: "HA4T", sym: "$8" },
+  { id: "HAWK", sym: "$6" },
+  { id: "HMNY", sym: "$1" },
+  { id: "HUCO", sym: "$7" },
+  { id: "HUNT", sym: "$6" },
+  { id: "HURI", sym: "$1" },
+  { id: "HX2", sym: "$7" },
+  { id: "IL76", sym: "$5" },
+  { id: "IL96", sym: "$5" },
+  { id: "IS2", sym: "$7" },
+  { id: "J328", sym: "$D" },
+  { id: "JAB4", sym: "$1" },
+  { id: "JAG2", sym: "$7" },
+  { id: "JPRO", sym: "$6" },
+  { id: "JS31", sym: "$2" },
+  { id: "JS32", sym: "$2" },
+  { id: "JS41", sym: "$2" },
+  { id: "JS42", sym: "$2" },
+  { id: "K126", sym: "$7" },
+  { id: "K226", sym: "$7" },
+  { id: "K35", sym: "$5" },
+  { id: "KA25", sym: "$7" },
+  { id: "KA26", sym: "$7" },
+  { id: "KA27", sym: "$7" },
+  { id: "KA50", sym: "$7" },
+  { id: "KA52", sym: "$7" },
+  { id: "KA62", sym: "$7" },
+  { id: "KC10", sym: "$4" },
+  { id: "KH4", sym: "$7" },
+  { id: "KMAX", sym: "$7" },
+  { id: "L101", sym: "$4" },
+  { id: "L410", sym: "$2" },
+  { id: "LAMA", sym: "$7" },
+  { id: "LANC", sym: "$C" },
+  { id: "LJ25", sym: "$8" },
+  { id: "LJ31", sym: "$8" },
+  { id: "LJ35", sym: "$8" },
+  { id: "LJ36", sym: "$8" },
+  { id: "LJ40", sym: "$8" },
+  { id: "LJ45", sym: "$8" },
+  { id: "LJ55", sym: "$8" },
+  { id: "LJ60", sym: "$8" },
+  { id: "LR2T", sym: "$7" },
+  { id: "LYNX", sym: "$7" },
+  { id: "M20", sym: "$1" },
+  { id: "M74", sym: "$7" },
+  { id: "MD11", sym: "$4" },
+  { id: "MD52", sym: "$7" },
+  { id: "MD60", sym: "$7" },
+  { id: "MD81", sym: "$3" },
+  { id: "MD82", sym: "$3" },
+  { id: "MD83", sym: "$3" },
+  { id: "MD87", sym: "$3" },
+  { id: "MD88", sym: "$3" },
+  { id: "MD90", sym: "$3" },
+  { id: "MH20", sym: "$7" },
+  { id: "MI10", sym: "$7" },
+  { id: "MI14", sym: "$7" },
+  { id: "MI2", sym: "$7" },
+  { id: "MI24", sym: "$7" },
+  { id: "MI26", sym: "$7" },
+  { id: "MI28", sym: "$7" },
+  { id: "MI34", sym: "$7" },
+  { id: "MI38", sym: "$7" },
+  { id: "MI4", sym: "$7" },
+  { id: "MI6", sym: "$7" },
+  { id: "MI8", sym: "$7" },
+  { id: "NA40", sym: "$7" },
+  { id: "NH90", sym: "$7" },
+  { id: "OH1", sym: "$7" },
+  { id: "P180", sym: "$2" },
+  { id: "P210", sym: "$1" },
+  { id: "P28A", sym: "$1" },
+  { id: "P28R", sym: "$1" },
+  { id: "P28T", sym: "$1" },
+  { id: "P32R", sym: "$1" },
+  { id: "P46T", sym: "$1" },
+  { id: "P68", sym: "$2" },
+  { id: "P750", sym: "$1" },
+  { id: "PA18", sym: "$1" },
+  { id: "PA22", sym: "$1" },
+  { id: "PA23", sym: "$2" },
+  { id: "PA24", sym: "$1" },
+  { id: "PA27", sym: "$2" },
+  { id: "PA28", sym: "$1" },
+  { id: "PA3", sym: "$2" },
+  { id: "PA4", sym: "$2" },
+  { id: "PAY3", sym: "$2" },
+  { id: "PC12", sym: "$1" },
+  { id: "PC6T", sym: "$1" },
+  { id: "PHIL", sym: "$7" },
+  { id: "PNR3", sym: "$1" },
+  { id: "PREM", sym: "$8" },
+  { id: "PRM1", sym: "$8" },
+  { id: "PSW4", sym: "$7" },
+  { id: "PTS1", sym: "$1" },
+  { id: "PTS2", sym: "$1" },
+  { id: "PUMA", sym: "$E" },
+  { id: "PUP", sym: "$1" },
+  { id: "R13", sym: "$5" },
+  { id: "R200", sym: "$1" },
+  { id: "R22", sym: "$7" },
+  { id: "R300", sym: "$1" },
+  { id: "R4", sym: "$7" },
+  { id: "R44", sym: "$7" },
+  { id: "R5", sym: "$1" },
+  { id: "R66", sym: "$7" },
+  { id: "RALL", sym: "$1" },
+  { id: "RJ1H", sym: "$9" },
+  { id: "RJ70", sym: "$9" },
+  { id: "RJ85", sym: "$9" },
+  { id: "RMOU", sym: "$7" },
+  { id: "RP1", sym: "$7" },
+  { id: "RV", sym: "$1" },
+  { id: "RVAL", sym: "$7" },
+  { id: "S22", sym: "$1" },
+  { id: "S274", sym: "$7" },
+  { id: "S278", sym: "$7" },
+  { id: "S330", sym: "$7" },
+  { id: "S360", sym: "$7" },
+  { id: "S51", sym: "$7" },
+  { id: "S52", sym: "$7" },
+  { id: "S55P", sym: "$7" },
+  { id: "S55T", sym: "$7" },
+  { id: "S58P", sym: "$7" },
+  { id: "S58T", sym: "$7" },
+  { id: "S61", sym: "$E" },
+  { id: "S61R", sym: "$E" },
+  { id: "S62", sym: "$E" },
+  { id: "S64", sym: "$E" },
+  { id: "S65C", sym: "$E" },
+  { id: "S76", sym: "$7" },
+  { id: "S92", sym: "$7" },
+  { id: "S92", sym: "$E" },
+  { id: "SB20", sym: "$D" },
+  { id: "SCOR", sym: "$7" },
+  { id: "SCOU", sym: "$7" },
+  { id: "SF34", sym: "$2" },
+  { id: "SH36", sym: "$2" },
+  { id: "SH4", sym: "$7" },
+  { id: "SIRA", sym: "$1" },
+  // { id: "SOL2", sym: "$SI2" },
+  { id: "SPIT", sym: "$1" },
+  { id: "SR20", sym: "$1" },
+  { id: "SR22", sym: "$1" },
+  { id: "SUBA", sym: "$1" },
+  { id: "SUCO", sym: "$7" },
+  { id: "SW3", sym: "$2" },
+  { id: "SW4", sym: "$2" },
+  { id: "SYCA", sym: "$7" },
+  { id: "T134", sym: "$3" },
+  { id: "T154", sym: "$3" },
+  { id: "T204", sym: "$4" },
+  { id: "TAMP", sym: "$1" },
+  { id: "TBM7", sym: "$1" },
+  { id: "TBM8", sym: "$1" },
+  { id: "TBM9", sym: "$2" },
+  { id: "TEX2", sym: "$1" },
+  { id: "TIGR", sym: "$7" },
+  { id: "TL20", sym: "$1" },
+  { id: "TOBA", sym: "$1" },
+  // { id: "TOR", sym: "$TORN" },
+  // { id: "TORN", sym: "$TORN" },
+  { id: "TRIN", sym: "$1" },
+  { id: "TUCA", sym: "$1" },
+  { id: "UH1", sym: "$7" },
+  { id: "UH12", sym: "$7" },
+  { id: "UH1Y", sym: "$7" },
+  { id: "ULTS", sym: "$7" },
+  { id: "V22", sym: "$A" },
+  { id: "V500", sym: "$7" },
+  { id: "W3", sym: "$7" },
+  { id: "WASP", sym: "$7" },
+  { id: "WESX", sym: "$7" },
+  { id: "WG30", sym: "$7" },
+  { id: "X2", sym: "$7" },
+  { id: "X49", sym: "$7" },
+  { id: "YK40", sym: "$3" },
+  { id: "YK42", sym: "$3" },
+  { id: "YK50", sym: "$1" },
+  { id: "YK52", sym: "$1" },
+  { id: "YNHL", sym: "$7" },
+  { id: "ZA6", sym: "$7" }
+];
+
+/* INITIALIZE AC_SYMBOL_MAP
+ * Aircraft symbol map is sorted in reverse order as otherwise
+ * short plane names (e.g., C5) would be selected prior
+ * to longer names (e.g., C550) and so the wrong icon
+ * would be selected.
+ */
+function sortPlaneSymbolList(a, b) {
+  var al = a.id;
+  var bl = b.id;
+
+  if (al < bl) return 1;
+  if (al > bl) return -1;
+  return 0;
+}
+ac_symbol_map = ac_symbol_map.sort(sortPlaneSymbolList);
+
+// based on ac cat and type, returns according svg path
+export function get_image_object(ac_type) {
+  var retval = {};
+  Object.assign(retval, acImagePathsR["$A320"]);
+  // Assign aircraft symbol based on ICAO type code */
+  if (ac_type != null) {
+    for (var i = 0; i < ac_symbol_map.length; i++) {
+      var ac_symbol = ac_symbol_map[i];
+      if (ac_type.indexOf(ac_symbol.id) == 0) {
+        retval = {};
+        Object.assign(retval, acImagePathsR[ac_symbol.sym]); // copy symbol to new object
+        return retval;
+      }
+    }
+  }
+  return retval;
+}
+// migrated from v0.1: why?????
+(function () {
+  acImagePathsR["$PARAGLIDER"].scale = 0.05;
+  acImagePathsR["$BEACON"].scale = 0.2;
+  acImagePathsR["$E"].scale = 0.95;
+  acImagePathsR["$SF25"].scale = 1.5;
+  acImagePathsR["$SI2"].scale = 0.78;
+  acImagePathsR["$AWACS"].scale = 0.68;
+  acImagePathsR["$A320"].scale = 0.6;
+  acImagePathsR["$A330"].scale = 0.53;
+  acImagePathsR["$A340"].scale = 0.58;
+  acImagePathsR["$A380"].scale = 0.6;
+  acImagePathsR["$B737"].scale = 0.61;
+  acImagePathsR["$B747"].scale = 0.9;
+  acImagePathsR["$B777"].scale = 0.6;
+  acImagePathsR["$B787"].scale = 0.58;
+  acImagePathsR["$F15"].scale = 0.53;
+
+  acImagePathsR["$0"].ofY = 2.0;
+  acImagePathsR["$1"].ofX = 0.0;
+
+  acImagePathsR["$2"].ofY = -1.5;
+  acImagePathsR["$2"].ofX = 0.2;
+  acImagePathsR["$3"].ofX = -1.15;
+  acImagePathsR["$4"].ofX = 0.1;
+
+  acImagePathsR["$5"].ofX = -0.0;
+  acImagePathsR["$A"].ofX = -0.2;
+  acImagePathsR["$A2"].ofX = -0.6;
+  acImagePathsR["$B"].ofX = -0.6;
+  acImagePathsR["$AWACS"].ofX = -1.2;
+
+  acImagePathsR["$SF25"].ofX = -1.1;
+  acImagePathsR["$SF25"].ofY = 1.2;
+
+  acImagePathsR["$SI2"].ofX = -0.3;
+  acImagePathsR["$SI2"].ofY = 3.7;
+
+  acImagePathsR["$A320"].ofY = -0.9;
+  acImagePathsR["$A320"].ofX = -0.4;
+  acImagePathsR["$A330"].ofY = -1.0;
+  acImagePathsR["$A330"].ofX = -1.1;
+  acImagePathsR["$A340"].ofY = -0.5;
+  acImagePathsR["$A340"].ofX = -0.3;
+  acImagePathsR["$A380"].ofX = -0.2;
+  acImagePathsR["$A400"].ofX = -1.05;
+  acImagePathsR["$B737"].ofY = -0.9;
+  acImagePathsR["$B737"].ofX = -0.4;
+  acImagePathsR["$F15"].ofX = -0.25;
+})();
